requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-ckeditor==6.7.0
django-import-export==3.2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
hashID==3.1.4
hashids==1.3.1
idna==3.4
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
xlrd==2.0.1
xlwt==1.3.0
